pyproject.toml
[project]
name = "med-palm-2"
version = "0.1.0"
description = ""
authors = [{ name = "Bharat Sharma", email = "bharat8717sharma@gmail.com" }]
readme = "README.md"
requires-python = ">=3.9,<4.0"
dependencies = [
    "fastapi>=0.115.11,<0.116.0",
    "uvicorn>=0.34.0,<0.35.0",
    "google-cloud-aiplatform>=1.85.0,<2.0.0",
    "python-dotenv>=1.0.1,<2.0.0",
    "google-generativeai>=0.8.4,<0.9.0",
    "medpalm>=0.2.0,<0.3.0",
]

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
